+++ GLMakie/src/drawing_primitives.jl
@@ -645,11 +645,16 @@ end
 # el32convert doesn't copy for array of Float32
 # But we assume that xy_convert copies when we use it
 xy_convert(x::AbstractArray, n) = copy(x)
-xy_convert(x, n) = [LinRange(extrema(x)..., n + 1);]
+xy_convert(x::Makie.EndPoints, n) = [LinRange(extrema(x)..., n + 1);]
 
 function draw_atomic(screen::Screen, scene::Scene, plot::Heatmap)
+    t = Makie.transform_func_obs(plot)
+
+    if plot.x[] isa Makie.EndPoints && plot.y[] isa Makie.EndPoints && Makie.is_identity_transform(t[])
+        # Fast path for regular heatmaps
+        return draw_image(screen, scene, plot)
+    end
     return cached_robj!(screen, scene, plot) do gl_attributes
-        t = Makie.transform_func_obs(plot)
         mat = plot[3]
         space = plot.space # needs to happen before connect_camera! call
         xypos = lift(plot, pop!(gl_attributes, :f32c), t, plot.model, plot[1], plot[2], space) do f32c, t, model, x, y, space
@@ -658,7 +663,7 @@ function draw_atomic(screen::Screen, scene::Scene, plot::Heatmap)
             # - model matrices with rotation & Float32 precisionissues
             x1d = xy_convert(x, size(mat[], 1))
             y1d = xy_convert(y, size(mat[], 2))
-            
+
             x1d = Makie.apply_transform_and_f32_conversion(f32c, t, model, x1d, 1, space)
             y1d = Makie.apply_transform_and_f32_conversion(f32c, t, model, y1d, 2, space)
 
@@ -685,12 +690,12 @@ function draw_atomic(screen::Screen, scene::Scene, plot::Heatmap)
     end
 end
 
-function draw_atomic(screen::Screen, scene::Scene, plot::Image)
+function draw_image(screen::Screen, scene::Scene, plot::Union{Heatmap, Image})
     return cached_robj!(screen, scene, plot) do gl_attributes
         position = lift(plot, plot[1], plot[2]) do x, y
             xmin, xmax = extrema(x)
             ymin, ymax = extrema(y)
-            rect =  Rect2(xmin, ymin, xmax - xmin, ymax - ymin)
+            rect = Rect2(xmin, ymin, xmax - xmin, ymax - ymin)
             return decompose(Point2d, rect)
         end
         gl_attributes[:vertices] = apply_transform_and_f32_conversion(plot, pop!(gl_attributes, :f32c), position)
@@ -707,6 +712,10 @@ function draw_atomic(screen::Screen, scene::Scene, plot::Image)
         end
         return draw_mesh(screen, gl_attributes)
     end
+end
+
+function draw_atomic(screen::Screen, scene::Scene, plot::Image)
+    return draw_image(screen, scene, plot)
 end
 
 function mesh_inner(screen::Screen, mesh, transfunc, gl_attributes, plot, space=:data)
@@ -879,7 +888,7 @@ function draw_atomic(screen::Screen, scene::Scene, plot::Volume)
         gl_attributes[:num_clip_planes] = Observable(0)
         pop!(gl_attributes, :clip_planes)
         gl_attributes[:clip_planes] = map(plot, gl_attributes[:modelinv], plot.clip_planes, plot.space) do modelinv, planes, space
-            Makie.is_data_space(space) || return [Vec4f(0, 0, 0, -1e9) for _ in 1:8] 
+            Makie.is_data_space(space) || return [Vec4f(0, 0, 0, -1e9) for _ in 1:8]
 
             # model/modelinv has no perspective projection so we should be fine
             # with just applying it to the plane origin and transpose(inv(modelinv))
@@ -890,7 +899,7 @@ function draw_atomic(screen::Screen, scene::Scene, plot::Volume)
             for i in 1:min(length(planes), 8)
                 origin = modelinv * to_ndim(Point4f, planes[i].distance * planes[i].normal, 1)
                 normal = transpose(gl_attributes[:model][]) * to_ndim(Vec4f, planes[i].normal, 0)
-                distance = dot(Vec3f(origin[1], origin[2], origin[3]) / origin[4], 
+                distance = dot(Vec3f(origin[1], origin[2], origin[3]) / origin[4],
                     Vec3f(normal[1], normal[2], normal[3]))
                 output[i] = Vec4f(normal[1], normal[2], normal[3], distance)
             end
@@ -952,7 +961,7 @@ function draw_atomic(screen::Screen, scene::Scene, plot::Voxels)
         gl_attributes[:num_clip_planes] = Observable(0)
         pop!(gl_attributes, :clip_planes)
         gl_attributes[:clip_planes] = map(plot, gl_attributes[:model], plot.clip_planes, plot.space) do model, planes, space
-            Makie.is_data_space(space) || return [Vec4f(0, 0, 0, -1e9) for _ in 1:8] 
+            Makie.is_data_space(space) || return [Vec4f(0, 0, 0, -1e9) for _ in 1:8]
 
             # model/modelinv has no perspective projection so we should be fine
             # with just applying it to the plane origin and transpose(inv(modelinv))
@@ -964,7 +973,7 @@ function draw_atomic(screen::Screen, scene::Scene, plot::Voxels)
             for i in 1:min(length(planes), 8)
                 origin = modelinv * to_ndim(Point4f, planes[i].distance * planes[i].normal, 1)
                 normal = transpose(model) * to_ndim(Vec4f, planes[i].normal, 0)
-                distance = dot(Vec3f(origin[1], origin[2], origin[3]) / origin[4], 
+                distance = dot(Vec3f(origin[1], origin[2], origin[3]) / origin[4],
                     Vec3f(normal[1], normal[2], normal[3]))
                 output[i] = Vec4f(normal[1], normal[2], normal[3], distance)
             end