+++ WGLMakie/src/wglmakie.bundled.js
@@ -22413,6 +22413,7 @@ function deserialize_plot(scene, data) {
     mesh.frustumCulled = false;
     mesh.matrixAutoUpdate = false;
     mesh.plot_uuid = data.uuid;
+    mesh.renderOrder = data.zvalue;
     update_visible(data.visible.value);
     data.visible.on(update_visible);
     connect_uniforms(mesh, data.uniform_updater);